Resolve memory leak.
This commit isn't for the LFU.

Before this commit, there are so many memory leak in MQSim code.
For example, when delete upcasting variable they don't cast to parent type, and when cache hit, they don't delete the transaction existed in cache slot.
So I fixed it to run this program with low memory compared to program before.